# Change encoding list of NaiveDateTimes to return the map directly