Closes #280

Notes:
1. In the future I would also like to move the call to new term to a function.
2. I also believe quick wins could be made in the encoding function when casting datetimes into the DF, as right now we have to do the *1000 when matching, I think we could use the built in datetime iter Polars has if we change it in the `s_new_date64` function.

Benchmarks:

Before:
```
Operating System: macOS
CPU Information: Apple M1
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.13.4
Erlang 25.0.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 1 s
reduction time: 0 ns
parallel: 1
inputs: 100, 1k, 10k, 100k, 1 million, 10 million
Estimated total run time: 1.30 min

Benchmarking NaiveDateTime change with input 100 ...
Benchmarking NaiveDateTime change with input 1k ...
Benchmarking NaiveDateTime change with input 10k ...
Benchmarking NaiveDateTime change with input 100k ...
Benchmarking NaiveDateTime change with input 1 million ...
Benchmarking NaiveDateTime change with input 10 million ...

##### With input 100 #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        4.98 K      200.79 μs    ±60.69%      183.29 μs      505.78 μs

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change        21.15 KB

**All measurements for memory usage were the same**

##### With input 1k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        1.15 K      872.83 μs    ±59.40%      803.19 μs     2269.71 μs

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change       196.93 KB

**All measurements for memory usage were the same**

##### With input 10k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        124.26        8.05 ms    ±13.95%        7.83 ms       11.12 ms

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change         1.91 MB

**All measurements for memory usage were the same**

##### With input 100k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change         62.38       16.03 ms     ±6.00%       15.96 ms       17.42 ms

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change        19.07 MB

**All measurements for memory usage were the same**

##### With input 1 million #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change          0.82         1.22 s     ±1.33%         1.21 s         1.26 s

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change       190.74 MB

**All measurements for memory usage were the same**

##### With input 10 million #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        0.0405        24.70 s     ±0.00%        24.70 s        24.70 s
```

Now:
```

##### With input 100 #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change       12.28 K       81.42 μs    ±35.07%       82.71 μs      143.15 μs

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change        20.99 KB

**All measurements for memory usage were the same**

##### With input 1k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        5.09 K      196.57 μs    ±14.45%      185.08 μs      282.58 μs

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change       196.63 KB

**All measurements for memory usage were the same**

##### With input 10k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change        817.88        1.22 ms     ±4.51%        1.21 ms        1.29 ms

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change         1.91 MB

**All measurements for memory usage were the same**

##### With input 100k #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change         62.57       15.98 ms     ±3.02%       15.97 ms       16.96 ms

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change        19.07 MB

**All measurements for memory usage were the same**

##### With input 1 million #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change          6.69      149.38 ms    ±23.18%      130.59 ms      261.13 ms

Memory usage statistics:

Name                    Memory usage
NaiveDateTime change       188.63 MB

**All measurements for memory usage were the same**

##### With input 10 million #####
Name                           ips        average  deviation         median         99th %
NaiveDateTime change          0.32         3.09 s    ±11.35%         3.18 s         3.41 s
```