@joshuataylor 2022-07-05
Happy for this to be reviewed.

This has the biggest change for larger series, where the time to process is now ~3s instead of ~24s :D